+++ tests/testthat/test-rescaler.R
@@ -1,30 +1,30 @@
 
-pal_blues <- palette_bezier(mat_luv_blues)
+pal_blues <- pal_luv_bezier(mat_luv_blues, rescale_path = FALSE)
 
-test_that("rescaler_linear_input works", {
+test_that("rescaler_x works", {
 
   # need the "\\[" notation for regular-expression escape
   expect_error(
-    rescaler_linear_input(c(-1, 0)),
+    rescaler_x(c(-1, 0)),
     "range\\[1\\] not greater than"
   )
 
   expect_error(
-    rescaler_linear_input(c(2, 0)),
+    rescaler_x(c(2, 0)),
     "range\\[1\\] not less than"
   )
 
   expect_error(
-    rescaler_linear_input(c(0, -1)),
+    rescaler_x(c(0, -1)),
     "range\\[2\\] not greater than"
   )
 
   expect_error(
-    rescaler_linear_input(c(0, 2)),
+    rescaler_x(c(0, 2)),
     "range\\[2\\] not less than"
   )
 
-  rlin <- rescaler_linear_input(c(0.25, 0.75))
+  rlin <- rescaler_x(c(0.25, 0.75))
 
   expect_s3_class(rlin, "cpath_rescaler")
 
@@ -33,7 +33,7 @@ test_that("rescaler_linear_input works", {
     c(0.25, 0.5, 0.75)
   )
 
-  rlin_rev <- rescaler_linear_input(c(0.75, 0.25))
+  rlin_rev <- rescaler_x(c(0.75, 0.25))
 
   expect_identical(
     rlin_rev(c(0, 0.5, 1)),
@@ -48,7 +48,8 @@ test_that("f_root_luminance works", {
   expect_type(froot, "closure")
   expect_equal(
     froot(c(0, 0.5, 1)),
-    c(30, 0, -30)
+    c(30, 0, -30),
+    tolerance = 1.e-3
   )
 
 })
@@ -63,10 +64,10 @@ test_that("root_luminance works", {
 
 })
 
-test_that("rescaler_linear_luminance works", {
+test_that("rescaler_lum works", {
 
-  rlum <- rescaler_linear_luminance(c(25, 55), pal_blues)
-  rlum2 <- rescaler_linear_luminance(c(55, 85), pal_blues)
+  rlum <- rescaler_lum(c(25, 55), pal_blues)
+  rlum2 <- rescaler_lum(c(55, 85), pal_blues)
 
   expect_s3_class(rlum, "cpath_rescaler")