+++ NEWS.md
@@ -1,5 +1,21 @@
 # colorpath 0.0.0 (development version)
 
+* Refactored (#1, continued):
+
+  - added arguments `rescale_path` and `n` to `pal_luv_bezier()`, rescales the palette to be perceptually uniform (in LUV space) with respect to the input. 
+  - made `rescaler_bezier()` an internal function.
+  - renamed function `palette_bezier()` to `pal_luv_bezier()`; added argument 
+    `rescale_path`, defaults `TRUE`, to indicate to rescale the input according 
+    to the path length in LUV space.
+  - reanmed function `rescale_palette()` to `rescale_pal_luv()`.
+  - renamed function `rescaler_linear_luminance()` to `rescaler_lum()`.
+  - renamed function `rescaler_linear_input()` to `rescaler_x()`.
+  - renamed S3 class `cpath_palette_luv` to `cpath_pal_luv`.
+  - added function `as_pal_disc()` to coerce to a discrete-palette function.
+  - renamed `palette_hex()` to `as_pal_hex()`.
+  - renamed function `luv()` to `as_mat_luv()`.
+  - renamed argument `luv` to `mat_luv`.
+
 * Added function to convert LUV palette-function to hex-code palette-function, `palette_hex()`. (#14)
 
 * Added function to rescale a palette function, `rescale_palette()`. (#10)